chore: eliminar archivos y dependencias de Spotify API: /n- Eliminar cualquier archivo relacionado con API de Spotify. /n- Actualizar "requirements.txt". /n- Reacomodar archivo principal de Deezer y renombrar como "main.py". /n- Closes issue #14.
