+++ main.py
@@ -1,88 +1,101 @@
-import spotipy
-import datetime
-import csv
-from spotipy.oauth2 import SpotifyOAuth
-from config import SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET, SPOTIFY_REDIRECT_URI
+import requests
 
-
-def get_spotify_client():
-    """
-    Crea y retorna un cliente autenticado de Spotify.
-    """
-    scope = "user-top-read user-read-recently-played"
-
-    sp = spotipy.Spotify(
-        auth_manager=SpotifyOAuth(
-            client_id=SPOTIFY_CLIENT_ID,
-            client_secret=SPOTIFY_CLIENT_SECRET,
-            redirect_uri=SPOTIFY_REDIRECT_URI,
-            scope=scope,
-        )
-    )
-    return sp
-
-
-def get_top_n_mexico_weekly(sp, pl_id, top=5) -> list[dict]:
-    """
-    Obtiene un objeto que contiene el top n de canciones de una playlist de Spotify.
-
-    Parámetros:
-    sp -- Autenticación de Spotify.
-    pl_id -- ID de alguna playlist de Spotify.
-    top -- Default 5. Establece el número de canciones.
-    """
-    results = sp.playlist_tracks(playlist_id=pl_id, limit=top)
-    return results["items"]
-
-
-if __name__ == "__main__":
-    # Prueba de conexión
-    print("Conectando a Spotify...")
-    client_personal = get_spotify_client()
-    print("✓ Conexión exitosa!")
-
-    # Lista de Spotify de Diskover Co.
-    id_top_100_mexico_weekly_discover_co = "5KLKS1zjjeqSe6oRgsdUMb"
-    url_top_100_mexico_weekly_discover_co = (
-        "https://open.spotify.com/playlist/5KLKS1zjjeqSe6oRgsdUMb"
-    )
-
-    tracks_object = get_top_n_mexico_weekly(
-        client_personal, pl_id=id_top_100_mexico_weekly_discover_co, top=50
-    )
-
-    # Lista de canciones.
-    tracks_list = []
-    header = ["posición", "canción", "albúm", "artista(s)"]
-    tracks_list.append(header)
-
-    for i, track_object in enumerate(tracks_object):
-        posicion = i + 1
-        cancion = track_object["track"]["name"]
-        album = track_object["track"]["album"]["name"]
-
-        artists_objects = track_object["track"]["artists"]
-        artistas = [artist["name"] for artist in artists_objects]
-
-        track_row = [posicion, cancion, album, artistas]
-        tracks_list.append(track_row)
-
-    # Creación de CSV.
+def obtener_id_artista(nombre_artista):
+    """Busca el ID del artista en Deezer usando el nombre."""
+    url = "https://api.deezer.com/search/artist"
+    parametros = {'q': nombre_artista}
+    
     try:
-        print(
-            "\nGenerando CSV de Top 50 canciones de Spotify mas escuchadas semanalmente en México \npor Discover Co."
-        )
-        file_name = "top-50-mexico-weekly"
-        csv_creation_date_utc = datetime.datetime.now(tz=datetime.UTC).strftime(
-            "%Y-%m-%d-%H-%M%z"
-        )
-        utc_file_path = f"{file_name}({csv_creation_date_utc}).csv"
-
-        with open(file=utc_file_path, mode="w", newline="", encoding="utf-8") as file:
-            writer = csv.writer(file)
-            for track in tracks_list:
-                writer.writerow(track)
+        response = requests.get(url, params=parametros)
+        response.raise_for_status()
+        datos = response.json()
+        
+        if datos['data']:
+            return datos['data'][0]['id']
+        else:
+            print(f"Artista '{nombre_artista}' no encontrado.")
+            return None
     except Exception as e:
-        print(f"\nError: {e}")
-    else:
-        print(f"\nEl archivo '{utc_file_path}' ha sido creado.")
+        print(f"Error al buscar el artista: {e}")
+        return None
+
+def obtener_albumes(artist_id):
+    """Obtiene la lista de álbumes de un artista mediante su ID."""
+    url = f"https://api.deezer.com/artist/{artist_id}/albums"
+    
+    try:
+        response = requests.get(url)
+        response.raise_for_status()
+        datos = response.json()
+        
+        lista_albumes = []
+        if 'data' in datos:
+            for album in datos['data']:
+                lista_albumes.append({
+                    "id_album": album['id'],
+                    "titulo": album['title'],
+                    "fecha_lanzamiento": album.get('release_date', 'Desconocida')
+                })
+        return lista_albumes
+    except Exception as e:
+        print(f"Error al obtener álbumes: {e}")
+        return []
+
+def obtener_top_canciones(artist_id):
+    """Obtiene las 5 canciones más escuchadas de un artista."""
+    url = f"https://api.deezer.com/artist/{artist_id}/top"
+    parametros = {'limit': 5}
+    
+    try:
+        response = requests.get(url, params=parametros)
+        response.raise_for_status()
+        datos = response.json()
+        
+        lista_canciones = []
+        if 'data' in datos:
+            for track in datos['data']:
+                lista_canciones.append({
+                    "id_track": track['id'],
+                    "titulo": track['title'],
+                    "duracion_segundos": track['duration'],
+                    "reproducciones": track.get('rank', 'N/A')
+                })
+        return lista_canciones
+    except Exception as e:
+        print(f"Error al obtener las canciones: {e}")
+        return []
+
+# --- Bloque Principal Modificado ---
+if __name__ == "__main__":
+    # El ciclo se repetirá exactamente 10 veces
+    for i in range(1, 11):
+        print(f"\n==============================")
+        print(f"   BÚSQUEDA {i} DE 10")
+        print(f"==============================")
+        
+        nombre = input("Ingresa el nombre del artista que deseas buscar: ")
+        
+        print(f"\nProcesando información de: {nombre}...")
+        
+        # 1. Obtener ID
+        id_artista = obtener_id_artista(nombre)
+        
+        if id_artista:
+            print(f"ID del artista en Deezer: {id_artista}")
+            
+            # 2. Obtener Álbumes
+            albumes = obtener_albumes(id_artista)
+            print(f"\n--- Álbumes Encontrados: {len(albumes)} ---")
+            # Mostramos los 3 primeros
+            for j, album in enumerate(albumes[:3], 1):
+                print(f"{j}. {album['titulo']} | Lanzamiento: {album['fecha_lanzamiento']}")
+                
+            # 3. Obtener Top 5 canciones
+            canciones = obtener_top_canciones(id_artista)
+            print("\n--- Top 5 Canciones Más Escuchadas ---")
+            for j, cancion in enumerate(canciones, 1):
+                print(f"{j}. {cancion['titulo']} | Duración: {cancion['duracion_segundos']}s | Popularidad: {cancion['reproducciones']}")
+        
+        print("\n" + "-"*30)
+
+    print("\nHas completado las 10 búsquedas programadas.")
