+++ requirements.txt
@@ -1,6 +1,1 @@
-certifi==2026.4.22
-charset-normalizer==3.4.7
-idna==3.13
-python-dotenv==1.2.2
 requests==2.33.1
-urllib3==2.6.3
